refactor(delivery): couche Application — Actions extraites des contrôleurs (Closes #6898)
- CreateDeliveryAction : DeliveryController::store + insertDelivery (idempotence
  source contractuelle, référence DLV-YYYY-NNNNNN, course 23505) ;
- UpdateDeliveryStopStatusAction : DeliveryRiderController::status (verrou
  SELECT FOR UPDATE, RBAC par propriété, POD, événements de tracking) ;
- SummarizeDeliveryCodSettlementsAction : DeliveryCodSettlementController::report
  (agrégats attendus vs collectés par statut).

Comportement API strictement identique : rejeu source → 200, création → 201
(conservés via Delivery::wasRecentlyCreated). Couverture : suites Feature
Delivery existantes (DeliveryApiTest, DeliveryRiderApiTest,
DeliveryCodSettlementApiTest, DeliveryRouteApiTest). PHPStan strict 0 erreur,
Pint OK.
