+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- **refactor(delivery): couche Application — Actions extraites des contrôleurs (Closes #6898).** `DeliveryController::store` (idempotence source contractuelle, référence DLV, course 23505) → `CreateDeliveryAction` ; `DeliveryRiderController::status` (verrou `SELECT FOR UPDATE`, RBAC par propriété, POD, événements de tracking) → `UpdateDeliveryStopStatusAction` ; `DeliveryCodSettlementController::report` (agrégats attendus vs collectés) → `SummarizeDeliveryCodSettlementsAction`. Comportement API strictement identique (rejeu → 200, création → 201, conservés via `wasRecentlyCreated`) ; couverture par les suites Feature `Delivery*` existantes. Aucun changement de contrat API.
+
 - **docs(specs): 3 spécifications produits livrées (assistant IA BC-23, vitrine BC-27, catalogue B2B BC-28) — validées fondateur 2026-09-06 (Closes #6847, #6863, #6878).** `docs/specifications/SOLUTION_COMMAND_ASSISTANT_MCP.md` (BC-23 : commandes texte & voix des managers exécutées dans l'app, contrat d'outil déclaratif par BC façon MCP, middleware permissions serveur, confirmation des actions sensibles, audit « via assistant IA », drivers LLM/STT interchangeables — Groq+Whisper par défaut, fake pour tests ; programme #6846→#6861), `SOLUTION_SITE_VITRINE.md` (BC-27 : site vitrine entreprise 1-clic, moteur de sections + thèmes maison — pas de CMS embarqué, spike GrapeJS pour l'éditeur, public isolé 0 donnée interne ; programme #6862→#6876), `SOLUTION_CATALOGUE_B2B.md` (BC-28 : produits B2B des tenants aux acheteurs, catalogue public + demandes de devis → leads CRM BC-11, minor units, pas de paiement v1 ; programme #6877→#6891). Aucun changement de code applicatif.
 - **ops(backup): backup & restore drill fail-loud quand les secrets manquent (Part of #6836, porté depuis #6843).** `.github/workflows/database-backup.yml` : le no-op silencieux (`::notice::` + exit 0) est remplacé par un échec volontaire (`::error::` + exit 1) lorsque DATABASE_URL/BACKUP_S3_BUCKET/AWS_* sont absents — aucune sauvegarde n'existe tant que les secrets ne sont pas posés (Neon gratuit, pas de PITR), l'alerte doit déclencher l'action (Settings → Secrets → Actions). Restore drill : même règle. Aucun changement de comportement applicatif.
 - **fix(tests): Camera*Test — 7 erreurs PHPStan strict résolues (findOrFail/firstOrFail/fresh() ?? $instance) après re-clé des FQCN suite au déplacement Domain/Models (PR #6832).**
